+++ python/tskit/trees.py
@@ -5236,6 +5236,11 @@ class TreeSequence:
         check that the alleles result in a valid VCF---for example, it is possible
         to use the tab character as an allele, leading to a broken VCF.
 
+        The ID value in the output VCF file is the integer ID of the corresponding
+        :ref:`site <sec_site_table_definition>` (``site.id``). Subsequently,
+        These ID values can be utilized to match the contents of the VCF file
+        to the sites in the tree sequence object.
+
         The ``position_transform`` argument provides a way to flexibly translate
         the genomic location of sites in tskit to the appropriate value in VCF.
         There are two fundamental differences in the way that tskit and VCF define